requirements.txt
Django==2.2.6
django-material-admin==1.7.0
djangorestframework==3.10.3
djangorestframework-simplejwt==4.4.0
Pillow==6.2.0
requests==2.22.0
